\name{fAssets-package}


\alias{fAssets}


\docType{package}


\title{Assets Modelling}


\description{

    The Rmetrics "fAssets" package is a very powerful collection
    of functions to investigate and analyze data sets of financial
    assets from different points ov view.

}


\details{

    \tabular{ll}{
        Package:    \tab fAssets\cr
        Type:       \tab Package\cr
        Date:       \tab 2009\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2008 Diethelm Wuertz and Rmetrics Association\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }
    
}


\keyword{models}

