\name{fAssets-package}


\alias{fAssets}


\docType{package}


\title{Assets Modelling}


\description{

    The Rmetrics "fAssets" package is a  collection of functions to 
    manage, to investigate and to analyze data sets of financial assets 
    from different points of view.

}


\details{

    \tabular{ll}{
        Package:    \tab fAssets\cr
        Type:       \tab Package\cr
        Date:       \tab 2009\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2014 Diethelm Wuertz and Rmetrics Association\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }
    
}


\keyword{models}

